pagedIO patches for STFC - backward client compatibility for xrd543 client and return type error for asyncpagereader

+++ src/XrdCl/XrdClAsyncPageReader.hh
@@ -189,8 +189,8 @@ class AsyncPageReader
     //--------------------------------------------------------------------------
     inline static uint32_t CalcIOVSize( uint32_t dleft )
     {
-      uint32_t ret = ( dleft / PageWithDigest + 2 ) * 2;
-      return ( ret > uint32_t( max_iovcnt() ) ? max_iovcnt() : ret );
+      int ret = ( dleft / PageWithDigest + 2 ) * 2;
+      return ( ret > ( max_iovcnt() ) ? max_iovcnt() : ret );
     }
 
     //--------------------------------------------------------------------------
